Splitting ActivitySlot for timetable cells into TimeSlot and TimetableCell classes
Can now use TimeSlot in a more general sense, where it represents some time range on a particular day.

Also, TimetableCell is more generic in that it doesn't need to reference an activity - it could contain anything (e.g. maybe there will be more than just "activities" that can go in the timetable in the future.